requirements.txt
streamlit
langchain
pyodbc
python-dotenv
langchain-community
openai
psycopg2-binary==2.9.10
langchain-openai
pymssql
boto3
langchain-aws
pandas
matplotlib
seaborn
boto3
oracledb
snowflake-connector-python
snowflake-sqlalchemy
snowflake-connector-python[pandas]
